requirements.txt
google-generativeai
python-dotenv
tqdm
huggingface-hub
torch
accelerate
transformers
diffusers
numpy
scipy
librosa
ml-collections
absl-py
gradio
av
aiortc
pydub
wheel
librosa
openai-whisper
thefuzz
python-Levenshtein
nltk